feat(userinfo): update and fix userinfo
